Create initial Docker deployment configuration
I adapted the Docker and Docker Compose configuration files from the
official Docker Samples repository for Laravel:
https://github.com/dockersamples/laravel-docker-examples

Note: I added `-h postgres` to the healthcheck of the postgres service
as it brought up a FATAL error saying the role "postgres" did not exist.
Apparently, the issue had to do with Docker networking. See this
StackOverflow answer for more context:
https://stackoverflow.com/a/78477653